requirements.txt
# Core Flask (versiones compatibles entre sí)
Flask==1.1.4
Jinja2==2.11.3
MarkupSafe==2.0.1
Werkzeug==1.0.1
itsdangerous==1.1.0
click==7.1.2
setuptools

# Base de Datos
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.52
psycopg2-binary==2.9.9

# Serialización (versiones compatibles)
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
Flask-Marshmallow==0.14.0

# API REST y Seguridad
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.3.1
PyJWT==2.3.0

# Utilidades
python-dotenv==0.19.2
